+++ Example/FlaneurImagePicker/DemoImagePickerViewController.swift
@@ -1,16 +1,7 @@
-//
-//  ViewController.swift
-//  FlaneurImagePickerController
-//
-//  Created by Flâneur on 11/07/2017.
-//
-//
-
 import UIKit
 import FlaneurImagePicker
 
-
-class ViewController: UIViewController {
+class DemoImagePickerViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -23,10 +14,10 @@ class ViewController: UIViewController {
     }
 
     @IBAction func lauchPickerButtonTouched() {
-        let images =  [
-            FlaneurImageDescription(imageURLString: "https://cdn.pixabay.com/photo/2017/07/24/02/40/pink-roses-2533389_960_720.jpg")!,
-            FlaneurImageDescription(imageURLString: "https://cdn.pixabay.com/photo/2017/04/04/14/24/turtle-2201433_960_720.jpg")!,
-            FlaneurImageDescription(imageURLString: "https://cdn.pixabay.com/photo/2017/07/28/16/30/bee-pollen-2549125_960_720.jpg")!
+        let images: [FlaneurImageDescriptor] = [
+            .url(URL(string: "https://cdn.pixabay.com/photo/2017/07/24/02/40/pink-roses-2533389_960_720.jpg")!),
+            .url(URL(string: "https://cdn.pixabay.com/photo/2017/04/04/14/24/turtle-2201433_960_720.jpg")!),
+            .url(URL(string: "https://cdn.pixabay.com/photo/2017/07/28/16/30/bee-pollen-2549125_960_720.jpg")!),
         ]
 
         let flaneurPicker = FlaneurImagePickerController(userInfo: nil,
@@ -88,13 +79,13 @@ class ViewController: UIViewController {
 
 }
 
-extension ViewController: FlaneurImagePickerControllerDelegate {
+extension DemoImagePickerViewController: FlaneurImagePickerControllerDelegate {
     func flaneurImagePickerControllerDidCancel(_ picker: FlaneurImagePickerController) {
         print("didCancel")
         picker.dismiss(animated: true)
     }
 
-    func flaneurImagePickerController(_ picker: FlaneurImagePickerController, didFinishPickingImages images: [FlaneurImageDescription], userInfo: Any?) {
+    func flaneurImagePickerController(_ picker: FlaneurImagePickerController, didFinishPickingImages images: [FlaneurImageDescriptor], userInfo: Any?) {
         print("didFinishPickingImages")
         for image in images {
             print("* \(image)")
@@ -112,7 +103,7 @@ extension ViewController: FlaneurImagePickerControllerDelegate {
 
     func flaneurImagePickerController(_ picker: FlaneurImagePickerController,
                                       withCurrentSelectionOfSize count: Int,
-                                      actionForNewImageSelection: FlaneurImageDescription) ->
+                                      actionForNewImageSelection: FlaneurImageDescriptor) ->
         FlaneurImagePickerControllerAction {
             debugPrint("Count is \(count)")
             if count < 5 {