+++ Example/FlaneurImagePicker/ItemsTableViewController.swift
@@ -1,0 +1,59 @@
+import UIKit
+
+/// The descriptor for a table view cell.
+public struct TableViewCellDescriptor {
+    let cellClass: UITableViewCell.Type
+    let reuseIdentifier: String
+    let configure: (UITableViewCell) -> ()
+
+    public init<Cell: UITableViewCell>(reuseIdentifier: String, configure: @escaping (Cell) -> ()) {
+        self.cellClass = Cell.self
+        self.reuseIdentifier = reuseIdentifier
+        self.configure = { cell in
+            configure(cell as! Cell)
+        }
+    }
+}
+
+/// A generic implementation of a `UITableViewController`.
+///
+/// Based on [objc.io's Generic Table View Controllers](https://talk.objc.io/episodes/S01E26-generic-table-view-controllers-part-2).
+final class ItemsViewController<Item>: UITableViewController {
+    var items: [Item] = []
+    let cellDescriptor: (Item) -> TableViewCellDescriptor
+    public var didSelect: (Item) -> () = { _ in }
+    var reuseIdentifiers: Set<String> = []
+
+    public init(items: [Item], cellDescriptor: @escaping (Item) -> TableViewCellDescriptor) {
+        self.cellDescriptor = cellDescriptor
+        super.init(style: .plain)
+        self.items = items
+    }
+
+    required public init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override public func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        let item = items[indexPath.row]
+        didSelect(item)
+    }
+
+    override public func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return items.count
+    }
+
+    override public func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let item = items[indexPath.row]
+        let descriptor = cellDescriptor(item)
+
+        if !reuseIdentifiers.contains(descriptor.reuseIdentifier) {
+            tableView.register(descriptor.cellClass, forCellReuseIdentifier: descriptor.reuseIdentifier)
+            reuseIdentifiers.insert(descriptor.reuseIdentifier)
+        }
+
+        let cell = tableView.dequeueReusableCell(withIdentifier: descriptor.reuseIdentifier, for: indexPath)
+        descriptor.configure(cell)
+        return cell
+    }
+}